Add default variable env
The env variable will return a dictionary with all environmental
variables, it will be displayed as you would display dictionary with the
str built-in function. So some of the values might get truncated, but I
would like to add getitem functionality to the dictionary types
variables. Hence making it useful in the future.